test/smoke: move things around
