+++ test/smoke/build_splitcontainer_test.go
@@ -4,42 +4,7 @@ package smoke
 
 import (
 	"testing"
-
-	"github.com/opentable/sous/util/filemap"
 )
-
-func simpleServerSplitContainer() filemap.FileMap {
-	return filemap.FileMap{
-		"Dockerfile": `
-			FROM alpine:3.2
-			ENV SOUS_RUN_IMAGE_SPEC=/image-spec.json
-			COPY image-spec.json /
-			RUN mkdir /server
-			COPY server.sh /server/
-			`,
-		"image-spec.json": `
-			{
-			  "image": {
-			    "type": "Docker",
-				"from": "alpine:3.2"
-			  },
-			  "files": [
-			    {
-				  "source": {"dir": "/server"},
-			      "dest": {"dir": "/"}
-			    }
-			  ],
-			  "exec": ["/server/server.sh"]
-			}
-			`,
-		"server.sh": `#!/usr/bin/env sh
-			echo "Listening on :$PORT0"
-			while true; do
-			  echo -e "HTTP/1.1 200 OK\n\n$(date)" | nc -l -p $PORT0
-			done
-			`,
-	}
-}
 
 func TestSplitContainer(t *testing.T) {
